Rename linearmodels submodule to regression
